Shalev's_Projects/10_TmunoteAI/requirements.txt
torch
diffusers
transformers
accelerate
customtkinter
Pillow
flask
